07 - Reorganice

+++ 07 - Notes App (CRUD, Spring Security)/src/main/java/com/example/demo/controllers/UserController.java
@@ -1,9 +1,13 @@
-package com.example.demo;
+package com.example.demo.controllers;
 
+import com.example.demo.repositories.ListRepository;
+import com.example.demo.repositories.NotesRepository;
+import com.example.demo.repositories.UserRepository;
+import com.example.demo.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.Locale;
+import java.security.Principal;
 
 @RestController
 @RequestMapping("/user")
@@ -16,6 +20,7 @@ public class UserController {
     private ListRepository listRepository;
 
     //Users CRUD
+
     @PostMapping("/create")
     public String createUser(@RequestParam String username, @RequestParam String password){
         UserService userService = new UserService(username, password, listRepository, noteRepository, userRepository);
@@ -23,7 +28,10 @@ public class UserController {
     }
 
     @GetMapping
-    public String getUser(@RequestParam String username){
+    public String getUser(@RequestParam String username, Principal principal){
+        if(!principal.getName().equals(username)) {
+            return "No tienes acceso a este perfil";
+        }
         UserService userService = new UserService(username, listRepository, noteRepository, userRepository);
         return userService.getUser();
     }
@@ -35,7 +43,10 @@ public class UserController {
 //    }
 
     @DeleteMapping
-    public String deleteUser(@RequestParam String username, @RequestParam String password) {
+    public String deleteUser(@RequestParam String username, @RequestParam String password, Principal principal) {
+        if(!principal.getName().equals(username)) {
+            return "No puedes acceder a esta funcion";
+        }
         UserService userService = new UserService(username, password, listRepository, noteRepository, userRepository);
         return userService.deleteUser();
     }
